+++ protected/humhub/modules/space/modules/manage/views/member/change-owner.php
@@ -1,9 +1,14 @@
 <?php
 
+use humhub\modules\space\modules\manage\models\ChangeOwnerForm;
 use yii\helpers\Html;
 use humhub\modules\space\modules\manage\widgets\MemberMenu;
 use yii\widgets\ActiveForm;
+use humhub\widgets\Button;
+
+/* @var $model ChangeOwnerForm */
 ?>
+
 
 <div class="panel panel-default">
     <div class="panel-heading">
@@ -15,11 +20,12 @@ use yii\widgets\ActiveForm;
         <p><?= Yii::t('SpaceModule.manage', 'As owner of this space you can transfer this role to another administrator in space.'); ?></p>
 
         <?php $form = ActiveForm::begin([]); ?>
-        <?= $form->field($model, 'ownerId')->dropDownList($model->getNewOwnerArray()) ?>
 
-        <hr>
+            <?= $form->field($model, 'ownerId')->dropDownList($model->getNewOwnerArray()) ?>
 
-        <?= Html::submitButton(Yii::t('SpaceModule.manage', 'Transfer ownership'), ['class' => 'btn btn-danger', 'data-confirm' => 'Are you really sure?']) ?>
+            <hr>
+
+            <?= Button::danger(Yii::t('SpaceModule.manage', 'Transfer ownership'))->action('client.submit')->confirm() ?>
 
         <?php ActiveForm::end(); ?>