+++ protected/humhub/docs/CHANGELOG.md
@@ -9,6 +9,8 @@ HumHub Change Log
 - Fix: space archive activity wrong originator assignment
 - Fix: suppress "unable to determine dataType" error for aborted xhr requests
 - Enh: added `FunctionalTester::loginBySpaceUserGroup()` and `FunctionalTest::assertSpaceAccessStatus()` for ACL testing
+- Fix #2721 delete space button not visible for system admin
+- Enh: added `humhub\modules\space\behaviors\SpaceModelMembership::canDelete()`
 
 1.3.0-beta.2  (July 18, 2018)
 -----------------------------